Added conclusive collection of relevant scripts
If anything isn't working in here or is broken, it's either due to a specific requirement of reconfiguration with your system, or I've forgotten a very relevant and specific script that makes this machine run. Hopefully everything is well documented enough for this to be obvious upon any errors that occur. Sorry and thanks in advance!